+++ c7n/actions/__init__.py
@@ -20,4 +20,4 @@ from .metric import PutMetric
 from .network import ModifyVpcSecurityGroupsAction
 from .notify import BaseNotify, Notify
 from .policy import RemovePolicyBase, ModifyPolicyBase
-from .securityhub import OtherResourcePostFinding
+
