+++ c7n/filters/__init__.py
@@ -30,4 +30,3 @@ from .health import HealthEventFilter
 from .iamaccess import CrossAccountAccessFilter, PolicyChecker
 from .metrics import MetricsFilter, ShieldMetrics
 from .vpc import DefaultVpcBase
-from .securityhub import SecurityHubFindingFilter
