+++ c7n/policy.py
@@ -634,144 +634,6 @@ class EC2InstanceState(LambdaMode):
                      'stopped', 'stopping', 'terminated']}})
 
 
-@execution.register('hub-action')
-@execution.register('hub-finding')
-class SecurityHub(LambdaMode):
-    """
-    Execute a policy lambda in response to security hub finding event or action.
-
-    .. example:
-
-    This policy will provision a lambda and security hub custom action.
-    The action can be invoked on a finding or insight result (collection
-    of findings). The action name will have the resource type prefixed as
-    custodian actions are resource specific.
-
-    .. code-block: yaml
-
-       policy:
-         - name: remediate
-           resource: aws.ec2
-           mode:
-             type: hub-action
-             role: MyRole
-           actions:
-            - snapshot
-            - type: set-instance-profile
-              name: null
-            - stop
-
-    .. example:
-
-    This policy will provision a lambda that will process high alert findings from
-    guard duty (note custodian also has support for guard duty events directly).
-
-    .. code-block: yaml
-
-       policy:
-         - name: remediate
-           resource: aws.iam
-           filters:
-             - type: event
-               key: detail.findings[].ProductFields.aws/securityhub/ProductName
-               value: GuardDuty
-             - type: event
-               key: detail.findings[].ProductFields.aws/securityhub/ProductName
-               value: GuardDuty
-           actions:
-             - remove-keys
-
-    Note, for custodian we support additional resources in the finding via the Other resource,
-    so these modes work for resources that security hub doesn't natively support.
-
-    https://docs.aws.amazon.com/securityhub/latest/userguide/securityhub-cloudwatch-events.html
-    """
-
-    schema = utils.type_schema(
-        'hub-finding', aliases=('hub-action',),
-        rinherit=LambdaMode.schema)
-
-    ActionFinding = 'Security Hub Findings - Custom Action'
-    ActionInsight = 'Security Hub Insight Results'
-    ImportFinding = 'Security Hub Findings - Imported'
-
-    handlers = {
-        ActionFinding: 'resolve_action_finding',
-        ActionInsight: 'resolve_action_insight',
-        ImportFinding: 'resolve_import_finding'
-    }
-
-    def resolve_findings(self, findings):
-        rids = set()
-        for f in findings:
-            for r in f['Resources']:
-                # Security hub invented some new arn format for a few resources...
-                # detect that and normalize to something sane.
-                if r['Id'].startswith('AWS') and r['Type'] == 'AwsIamAccessKey':
-                    rids.add('arn:aws:iam::%s:user/%s' % (
-                        f['AwsAccountId'],
-                        r['Details']['AwsIamAccessKey']['UserName']))
-                elif not r['Id'].startswith('arn'):
-                    log.warning("security hub unknown id:%s rtype:%s",
-                                r['Id'], r['Type'])
-                else:
-                    rids.add(r['Id'])
-        return rids
-
-    def resolve_action_insight(self, event):
-        rtype = event['detail']['resultType']
-        rids = [list(i.keys())[0] for i in event['detail']['insightResults']]
-        client = utils.local_session(
-            self.policy.session_factory).client('securityhub')
-        insights = client.get_insights(
-            InsightArns=[event['detail']['insightArn']]).get(
-                'Insights', ())
-        if not insights or len(insights) > 1:
-            return []
-        insight = insights.pop()
-        params = {}
-        params['Filters'] = insight['Filters']
-        params['Filters'][rtype] = [
-            {'Comparison': 'EQUALS', 'Value': r} for r in rids]
-        findings = client.get_findings(**params).get('Findings', ())
-        return self.resolve_findings(findings)
-
-    def resolve_action_finding(self, event):
-        return self.resolve_findings(event['detail']['findings'])
-
-    def resolve_import_finding(self, event):
-        return self.resolve_findings(event['detail']['findings'])
-
-    def resolve_resources(self, event):
-        # For centralized setups in a hub aggregator account
-        self.assume_member(event)
-
-        event_type = event['detail-type']
-        arn_resolver = getattr(self, self.handlers[event_type])
-        arns = arn_resolver(event)
-
-        # Lazy import to avoid aws sdk runtime dep in core
-        from c7n.resources.aws import Arn
-        resource_map = {Arn.parse(r) for r in arns}
-
-        # sanity check on finding resources matching policy resource
-        # type's service.
-        if self.policy.resource_manager.type != 'account':
-            log.info(
-                "mode:security-hub resolve resources %s", list(resource_map))
-            if not resource_map:
-                return []
-            resource_arns = [
-                r for r in resource_map
-                if r.service == self.policy.resource_manager.resource_type.service]
-            resources = self.policy.resource_manager.get_resources(
-                [r.resource for r in resource_arns])
-        else:
-            resources = self.policy.resource_manager.get_resources([])
-            resources[0]['resource-arns'] = resource_arns
-        return resources
-
-
 @execution.register('asg-instance-state')
 class ASGInstanceState(LambdaMode):
     """a lambda policy that executes on an asg's ec2 instance state changes."""
